move app files to pkg/app
