+++ demo/cmd/demo-wasm/hello.go
@@ -3,7 +3,7 @@ package main
 import (
 	"syscall/js"
 
-	"github.com/maxence-charriere/app"
+	"github.com/maxence-charriere/app/pkg/app"
 )
 
 // Hello is a component that describes a hello world. It implements the
